IPEX Diffusers fix Model CPU Offload
